twin_scape/requirements.txt
django
django-unfold
dotenv
gunicorn
mysqlclient
whitenoise
pillow
django-storages
boto3
minio
celery
redis
requests
certifi
pip-system-certs